% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLGL.R
\name{bootstrapHclust}
\alias{bootstrapHclust}
\title{Hierarchical Clustering with distance matrix computed using bootstrap replicates}
\usage{
bootstrapHclust(X, frac = 1, B = 50, method = "ward.D2", nCore = NULL)
}
\arguments{
\item{X}{data}

\item{frac}{fraction of sample used at each replicate}

\item{B}{number of replicates}

\item{method}{desired method: "single", "complete", "average", "mcquitty", "ward.D", "ward.D2", "centroid", "median".}

\item{nCore}{number of cores}
}
\value{
An object of class \code{hclust}
}
\description{
Hierarchical Clustering with distance matrix computed using bootstrap replicates
}
\examples{
hc <- bootstrapHclust(USArrests, nCore = 1)
}
