% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{coef.cv.MLGL}
\alias{coef.cv.MLGL}
\title{Get coefficients from a \code{\link{cv.MLGL}} object}
\usage{
\method{coef}{cv.MLGL}(object, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{\code{\link{cv.MLGL}} object}

\item{s}{Either "lambda.1se" or "lambda.min"}

\item{...}{Not used. Other arguments to predict.}
}
\value{
A matrix with estimated coefficients for given values of s.
}
\description{
Get coefficients from a \code{\link{cv.MLGL}} object
}
\seealso{
\link{cv.MLGL}, \link{predict.cv.MLGL}
}
\author{
Quentin Grimonprez
}
