% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{coef.MLGL}
\alias{coef.MLGL}
\title{Get coefficients from a \code{\link{MLGL}} object}
\usage{
\method{coef}{MLGL}(object, s = NULL, ...)
}
\arguments{
\item{object}{\code{\link{MLGL}} object}

\item{s}{values of lambda. If NULL, use values from object}

\item{...}{Not used. Other arguments to predict.}
}
\value{
A matrix with estimated coefficients for given values of s.
}
\description{
Get coefficients from a \code{\link{MLGL}} object
}
\seealso{
\link{MLGL}, \link{predict.MLGL}
}
\author{
Quentin Grimonprez
}
