% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMT.R
\name{HMT}
\alias{HMT}
\title{Hierarchical Multiple Testing procedure

Apply Hierarchical Multiple Testing procedure on a \code{\link{MLGL}} object}
\usage{
HMT(res, X, y, control = c("FWER", "FDR"), alpha = 0.05,
  test = partialFtest)
}
\arguments{
\item{res}{\code{\link{MLGL}} object}

\item{X}{matrix of size n*p}

\item{y}{vector of size n. If loss = "logit", elements of y must be in {-1,1}}

\item{control}{either "FDR" or "FWER"}

\item{alpha}{control level for testing procedure}

\item{test}{test used in the testing procedure. Default is partialFtest for loss = "ls" and partialChisqTest for loss = "logit"}
}
\value{
a list containing :
\describe{
  \item{lambdaOpt}{lambda values maximizing the number of rejects}
  \item{var}{A vector containing the index of selected variables for the first \code{lambdaOpt} value}
  \item{group}{A vector containing the values index of selected groups for the first \code{lambdaOpt} value}
  \item{selectedGroups}{Selected groups for the first \code{lambdaOpt} value}
  \item{reject}{Selected groups for all lambda values}
  \item{alpha}{Control level}
  \item{test}{Test used in the testing procedure}
  \item{control}{"FDR" or "FWER"}
  \item{time}{Elapsed time}
}
}
\description{
Hierarchical Multiple Testing procedure

Apply Hierarchical Multiple Testing procedure on a \code{\link{MLGL}} object
}
\examples{
# least square loss
set.seed(42)
X <- simuBlockGaussian(50, 12, 5, 0.7)
y <- drop(X[,c(2,7,12)] \%*\% c(2,2,-2) + rnorm(50, 0, 0.5))
res <- MLGL(X, y)

out <- HMT(res, X, y)


}
\seealso{
\link{MLGL}
}
