% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.cv.HCgglasso}
\alias{plot.cv.HCgglasso}
\title{Plot the cross-validation obtained from \code{\link{cv.HCgglasso}} function}
\usage{
\method{plot}{cv.HCgglasso}(x, log.lambda = FALSE, ...)
}
\arguments{
\item{x}{\code{\link{cv.HCgglasso}} object}

\item{log.lambda}{If TRUE, use log(lambda) instead of lambda in abscissa}

\item{...}{Other parameters for plot function}
}
\description{
Plot the cross-validation obtained from \code{\link{cv.HCgglasso}} function
}
\examples{
\dontrun{
X=simuBlockGaussian(50,12,5,0.7)
y=drop(X[,c(2,7,12)]\%*\%c(2,2,-2)+rnorm(50,0,0.5))
res=cv.HCgglasso(X,y)
plot(res) 
}

}
\seealso{
\link{cv.HCgglasso}
}

