% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.stability.HCgglasso}
\alias{plot.stability.HCgglasso}
\title{Plot the stability path obtained from \code{\link{stability.HCgglasso}} function}
\usage{
\method{plot}{stability.HCgglasso}(x, log.lambda = FALSE, threshold = 0.75,
  ...)
}
\arguments{
\item{x}{\code{\link{stability.HCgglasso}} object}

\item{log.lambda}{If TRUE, use log(lambda) instead of lambda in abscissa}

\item{threshold}{Threshold for selection frequency}

\item{...}{Other parameters for plot function}
}
\value{
A list containing :
\describe{
\item{var}{Index of selected variables for the given threshold.}
\item{group}{Index of the associated group.}
\item{threshold}{Value of threshold}
}
}
\description{
Plot the stability path obtained from \code{\link{stability.HCgglasso}} function
}
\examples{
\dontrun{
X=simuBlockGaussian(50,12,5,0.7)
y=drop(X[,c(2,7,12)]\%*\%c(2,2,-2)+rnorm(50,0,0.5))
res=stability.HCgglasso(X,y)
selected=plot(res)
print(selected)
}

}
\seealso{
\link{stability.HCgglasso}
}

