% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.HMT}
\alias{print.HMT}
\title{Print Values

Print a \code{\link{HMT}} object}
\usage{
\method{print}{HMT}(x, ...)
}
\arguments{
\item{x}{\code{\link{HMT}} object}

\item{...}{Not used.}
}
\description{
Print Values

Print a \code{\link{HMT}} object
}
\examples{
set.seed(42)
# Simulate gaussian data with block-diagonal variance matrix containing 12 blocks of size 5
X <- simuBlockGaussian(50, 12, 5, 0.7)
# Generate a response variable
y <- X[,c(2,7,12)]\%*\%c(2,2,-2) + rnorm(50,0,0.5)
# Apply MLGL method
res <- MLGL(X, y)
out <- HMT(res, X, y)
print(out)

}
\seealso{
\link{HMT} \link{summary.HMT}
}
