% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.MLGL}
\alias{print.MLGL}
\title{Print Values

Print a \code{\link{MLGL}} object}
\usage{
\method{print}{MLGL}(x, ...)
}
\arguments{
\item{x}{\code{\link{MLGL}} object}

\item{...}{Not used.}
}
\description{
Print Values

Print a \code{\link{MLGL}} object
}
\seealso{
\link{MLGL} \link{summary.MLGL}
}
