% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuBlockGaussian.R
\name{simuBlockGaussian}
\alias{simuBlockGaussian}
\title{Simulate multivariate Gaussian samples with block diagonal varaince matrix}
\usage{
simuBlockGaussian(n, nBlock, sizeBlock, rho)
}
\arguments{
\item{n}{number of samples to simulate}

\item{nBlock}{number of blocks}

\item{sizeBlock}{size of blocks}

\item{rho}{correlation within each block}
}
\value{
a matrix of size n *(nBlock*sizeBlock) containing the samples
}
\description{
simulate n samples from a gaussian multivariate law with 0 vector mean and block diagonal variance matrix with diagonal 1 and block of rho.
}
\examples{
X <- simuBlockGaussian(50, 12, 5, 0.7)
}
\author{
Quentin Grimonprez
}
