% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.fullProcess}
\alias{summary.fullProcess}
\title{Object Summaries

Summary of a \code{\link{fullProcess}} object}
\usage{
\method{summary}{fullProcess}(object, ...)
}
\arguments{
\item{object}{\code{\link{fullProcess}} object}

\item{...}{Not used.}
}
\value{
A matrix with fitted values or estimated coefficients for given values of s.
}
\description{
Object Summaries

Summary of a \code{\link{fullProcess}} object
}
\seealso{
\link{fullProcess} \link{print.fullProcess}
}
