% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.MLGL}
\alias{summary.MLGL}
\title{Object Summaries

Summary of a \code{\link{MLGL}} object}
\usage{
\method{summary}{MLGL}(object, ...)
}
\arguments{
\item{object}{\code{\link{MLGL}} object}

\item{...}{Not used.}
}
\value{
A matrix with fitted values or estimated coefficients for given values of s.
}
\description{
Object Summaries

Summary of a \code{\link{MLGL}} object
}
\seealso{
\link{MLGL} \link{print.MLGL}
}
