% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueGroupHclust.R
\name{uniqueGroupHclust}
\alias{uniqueGroupHclust}
\title{Find all unique groups in \code{\link{hclust}} results}
\usage{
uniqueGroupHclust(hc)
}
\arguments{
\item{hc}{output of \code{\link{hclust}} function}
}
\value{
A list containing:
\describe{
  \item{indexGroup}{Vector containing the index of variables.}
  \item{varGroup}{Vector containing the index of the group of each variable.}
}
}
\description{
Find all unique groups in \code{\link{hclust}} results
}
\examples{
hc <- hclust(dist(USArrests), "average")
res <- uniqueGroupHclust(hc)
}
\author{
Quentin Grimonprez
}
