% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.MLGL}
\alias{predict.MLGL}
\title{Predict fitted values from a \code{\link{MLGL}} object}
\usage{
\method{predict}{MLGL}(object, newx = NULL, s = NULL, type = c("fit", "coefficients"), ...)
}
\arguments{
\item{object}{\code{\link{MLGL}} object}

\item{newx}{matrix with new individuals for prediction. If type="coefficients", the parameter has to be NULL}

\item{s}{values of lambda. If NULL, use values from object}

\item{type}{if "fit", return the fitted values for each values of s, if "coefficients", return the estimated coefficients for each s}

\item{...}{Not used. Other arguments to predict.}
}
\value{
A matrix with fitted values or estimated coefficients for given values of s.
}
\description{
Predict fitted values from a \code{\link{MLGL}} object
}
\seealso{
\link{MLGL}
}
\author{
original code from \pkg{gglasso} package Author: Yi Yang <yiyang@umn.edu>, Hui Zou <hzou@stat.umn.edu>

function inspired from predict function from gglasso package by Yi Yang and Hui Zou.
}
